fix: use correct artifact name when publishing an sbt plugin (#391)
I suspect this was always wrong, but it's actually causing our
publication process to fail now that we've migrated to publishing
via the Maven Central Portal.

I'm not entirely confident about the approach, but it seems like at
least an improvement over the status quo, and we can further refine
the logic (and/or more properly reuse existing sbt code) later.